refactor: 필드 final 선언 및 패키지 분리
