+++ src/main/java/roomescape/domain/ReservationTime/ReservationTime.java
@@ -1,15 +1,12 @@
-package roomescape.domain;
+package roomescape.domain.ReservationTime;
 
 import java.time.LocalTime;
 import java.util.Objects;
 
 public class ReservationTime {
 
-    private Long id;
-    private LocalTime startAt;
-
-    public ReservationTime() {
-    }
+    private final Long id;
+    private final LocalTime startAt;
 
     public ReservationTime(Long id, LocalTime startAt) {
         this.id = id;
