+++ src/main/java/roomescape/dto/ReservationTimeRequest.java
@@ -3,7 +3,7 @@ package roomescape.dto;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
 import java.util.Objects;
-import roomescape.domain.ReservationTime;
+import roomescape.domain.ReservationTime.ReservationTime;
 
 public record ReservationTimeRequest(
         @JsonFormat(pattern = "HH:mm")
