+++ src/main/java/roomescape/repository/JdbcThemeRepository.java
@@ -8,7 +8,7 @@ import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.Theme;
+import roomescape.domain.Theme.Theme;
 
 @Repository
 public class JdbcThemeRepository implements ThemeRepository {
