+++ src/main/java/roomescape/repository/ReservationRepository.java
@@ -2,7 +2,7 @@ package roomescape.repository;
 
 import java.time.LocalDate;
 import java.util.List;
-import roomescape.domain.Reservation;
+import roomescape.domain.Reservation.Reservation;
 
 public interface ReservationRepository {
